+++ src/scandeval/config.py
@@ -123,6 +123,15 @@ class BenchmarkConfig:
         openai_api_key:
             The API key for the OpenAI API. If None then OpenAI models will not be
             benchmarked.
+        azure_openai_api_key:
+            The API key for the Azure OpenAI API. If None then Azure OpenAI models will
+            not be benchmarked.
+        azure_openai_endpoint:
+            The endpoint for the Azure OpenAI API. If None then Azure OpenAI models will
+            not be benchmarked.
+        azure_openai_api_version:
+            The api version for the Azure OpenAI API, e.g. "2023-12-01-preview". If
+            None then Azure OpenAI models will not be benchmarked.
         force:
             Whether to force the benchmark to run even if the results are already
             cached.
@@ -165,6 +174,9 @@ class BenchmarkConfig:
     evaluate_train: bool
     token: bool | str | None
     openai_api_key: str | None
+    azure_openai_api_key: str | None
+    azure_openai_endpoint: str | None
+    azure_openai_api_version: str | None
     force: bool
     progress_bar: bool
     save_results: bool