+++ tests/test_cli.py
@@ -33,6 +33,11 @@ def test_cli_param_names(params):
         "cache_dir",
         "token",
         "use_token",
+        "openai_api_key",
+        "prefer_azure",
+        "azure_openai_api_key",
+        "azure_openai_endpoint",
+        "azure_openai_api_version",
         "force",
         "framework",
         "device",
@@ -64,6 +69,11 @@ def test_cli_param_types(params):
     assert params["cache_dir"] == STRING
     assert params["token"] == STRING
     assert params["use_token"] == BOOL
+    assert params["openai_api_key"] == STRING
+    assert params["prefer_azure"] == BOOL
+    assert params["azure_openai_api_key"] == STRING
+    assert params["azure_openai_endpoint"] == STRING
+    assert params["azure_openai_api_version"] == STRING
     assert params["force"] == BOOL
     assert isinstance(params["framework"], Choice)
     assert isinstance(params["device"], Choice)