+++ CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+- Support for Azure OpenAI models! These can now be benchmarked as with any other
+  model, where either the environment variables `AZURE_OPENAI_API_KEY`,
+  `AZURE_OPENAI_ENDPOINT` and `AZURE_OPENAI_API_VERSION` need to have been set, or
+  alternatively through the `--azure-openai-api-key`, `--azure-openai-endpoint` and
+  `--azure-openai-api-version` arguments.
+- We now use the new JSON mode for newer OpenAI models for the NER task, to ensure
+  better JSON generation.
+- If an error is thrown during generation with an OpenAI model, which for instance
+  happens when the prompt is caught by the content filter, then we simply return a
+  blank string instead.
+
+
 ## [v12.3.2] - 2024-03-19
 ### Fixed
 - There is an issue with the underlying `outlines` package that we use for structured